refactor: isolate marble editor, game and wizard files into subfolders
Files used only by the build editor move to editor/, race-only files to game/,
and the lobby wizard to wizard/; shared track geometry, config, types, bedroom
and session stay at the root. Imports and a doc source-file path updated
accordingly. No behaviour change.
