+++ src/views/Games/MarbleEditor/MarbleEditor.vue
@@ -28,8 +28,8 @@ import { isMobile } from '@webgamekit/controls'
 import TouchControl from '@/components/TouchControl.vue'
 import { useMenuFocus } from '@/composables/useMenuFocus'
 import { EDITOR_MENU_MAPPING } from './config'
-import { useMarbleEditor } from './useMarbleEditor'
-import { useMarbleRace } from './useMarbleRace'
+import { useMarbleEditor } from './editor/useMarbleEditor'
+import { useMarbleRace } from './game/useMarbleRace'
 import { useMarbleEditorSession } from './useMarbleEditorSession'
 import EditorPalette from './EditorPalette.vue'
 import EditorToolbar from './EditorToolbar.vue'
